+++ mrv2/lib/mrvPanels/mrvComparePanel.h
@@ -19,6 +19,7 @@ namespace mrv
 
     class ClipButton;
     class HorSlider;
+    class PopupMenu;
 
     namespace panel
     {
@@ -31,6 +32,7 @@ namespace mrv
             HorSlider* wipeY;
             HorSlider* wipeRotation;
             HorSlider* overlay;
+            PopupMenu* compareTimeW;
 
         public:
             ComparePanel(ViewerUI* ui);
@@ -41,6 +43,7 @@ namespace mrv
             void redraw();
 
             void setCompareOptions(const tl::timeline::CompareOptions&);
+            void setCompareTime(const timeline::CompareTimeMode&);
 
             void refresh();